separate save and preprocessing functions from general.py
